added a 404 page; a few renamings
