+++ src/App.js
@@ -1,6 +1,6 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
 import ProtectedRoute from './Components/ProtectedRoute';
-import Home from './Pages/Home';
+import Notes from './Pages/Notes';
 import SingleNote from './Pages/SingleNote';
 import Authentication from './Pages/Authentication';
 import Layout from './Components/Layout';
@@ -15,14 +15,14 @@ function App() {
         <Route element={<ProtectedRoute />}>
           <Route path="/" element={<Layout />}>
             <Route index element={<Navigate to="/notes" />} />
-            <Route path="notes" element={<Home />} />
+            <Route path="notes" element={<Notes />} />
             <Route path="notes/:noteId" element={<SingleNote />} />
             <Route path="create" element={<CreateNote />} />
             <Route path="settings" element={<Settings />} />
             <Route path="auth" element={<Authentication />} />
-            <Route path="*" element={<PageNotFound />} />
           </Route>
         </Route>
+        <Route path="*" element={<PageNotFound />} />
       </Routes>
     </div>
   );
